bed working

+++ limix_reader/reader/plink/cbed/impl.c
@@ -9,38 +9,38 @@ inline double _long2double(char item)
     return item;
 }
 
-inline void _inv_strides(int* strides)
+inline void _inv_strides(long* strides)
 {
-    int tmp = strides[0];
+    long tmp = strides[0];
     strides[0] = strides[1];
     strides[1] = tmp;
 }
 
-void _read_intidx(FILE* fp, int* shape,
-                  int rn, long* rows, int cn, long* cols,
-                  double* matrix, int* strides)
+void _read_intidx(FILE* fp, long* shape,
+                  long rn, long* rows, long cn, long* cols,
+                  double* matrix, long* strides)
 {
-    int r, c;
+    long r, c;
     ItemIdx idx;
-    int index;
+    long index;
     for (r = 0; r < rn; r++)
     {
         for (c = 0; c < cn; c++)
         {
             idx.r = rows[r];
             idx.c = cols[c];
-            index = r * (strides[0]/sizeof(long))
-                  + c * (strides[1]/sizeof(long));
+            index = r * (strides[0]/sizeof(double))
+                  + c * (strides[1]/sizeof(double));
             matrix[index] = _long2double(read_item(fp, shape, &idx));
         }
     }
 }
 
-void read_intidx(FILE* fp, int* shape,
-                  int rn, long* rows, int cn, long* cols,
-                  double* matrix, int* strides, int same_axis)
+void read_intidx(FILE* fp, long* shape,
+                  long rn, long* rows, long cn, long* cols,
+                  double* matrix, long* strides, long same_axis)
 {
-    int _strides[2] = {strides[0], strides[1]};
+    long _strides[2] = {strides[0], strides[1]};
 
     if (same_axis)
         _read_intidx(fp, shape, rn, rows, cn, cols,
