+++ limix_reader/reader/plink/cbed/interface.h
@@ -1,10 +1,10 @@
 #ifndef INTERFACE_H
 #define INTERFACE_H
 
-int bed_item(char* filepath, int nsamples, int nmarkers, int sample, int marker);
+long bed_item(char* filepath, long nsamples, long nmarkers, long sample, long marker);
 
-void bed_intidx(char* filepath, int nsamples, int nmarkers,
-                int sn, long* samples, int mn, long* markers, double* matrix,
-                int* strides);
+void bed_intidx(char* filepath, long nsamples, long nmarkers,
+                long sn, long* samples, long mn, long* markers, double* matrix,
+                long* strides);
 
 #endif
